Moved registration of defaults into methods
